feat: Move ItemFinder and LineFinder to separate modals
