+++ app/javascript/src/components/editor/Modals/Modal.svelte
@@ -5,10 +5,14 @@
 
   export let align = "top"
   export let flush = false
+  export let transparent = false
+  export let maxWidth = null
 
   function close() {
     modal.close()
   }
+
+  $: console.log($modal)
 </script>
 
 <div
@@ -17,7 +21,7 @@
   use:escapeable on:escape={() => modal.close()}
   data-ignore>
 
-  <div class="modal__content" class:p-0={flush}>
+  <div class="modal__content" class:bg-transparent={transparent} class:p-0={flush} style:max-width={maxWidth}>
     <slot {close} />
   </div>
