+++ app/javascript/src/components/editor/Modals/Modals.svelte
@@ -1,12 +1,16 @@
 <script>
-  import BackupsModal from "../Modals/BackupsModal.svelte"
-  import CreateProjectModal from "../Modals/CreateProjectModal.svelte"
+  import BackupsModal from "./BackupsModal.svelte"
+  import CreateProjectModal from "./CreateProjectModal.svelte"
+  import LineFinderModal from "./LineFinderModal.svelte"
+  import ItemFinderModal from "./ItemFinderModal.svelte"
   import { modal } from "../../../stores/editor"
   import { Modal } from "../../../constants/Modal"
 
   const components = {
     [Modal.Backups]: BackupsModal,
-    [Modal.CreateProject]: CreateProjectModal
+    [Modal.CreateProject]: CreateProjectModal,
+    [Modal.LineFinder]: LineFinderModal,
+    [Modal.ItemFinder]: ItemFinderModal
   }
 
   $: component = components[$modal?.key]
