+++ app/javascript/src/constants/Modal.js
@@ -1,4 +1,6 @@
 export const Modal = Object.freeze({
   Backups: "backups",
-  CreateProject: "create-project"
+  CreateProject: "create-project",
+  LineFinder: "line-finder",
+  ItemFinder: "item-finder"
 })
